Implement relocations aswell

+++ api/src/main/java/com/artillexstudios/axapi/libraries/LibraryDownloader.java
@@ -11,17 +11,21 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 public final class LibraryDownloader {
     private final LibraryCache libraryCache;
     private final TransitiveDependencyCollector collector;
+    private final RelocationHelper helper;
     private final List<Library> libraries = new ArrayList<>();
     private final List<Repository> repositories = new ArrayList<>();
     private final List<Relocation> relocations = new ArrayList<>();
     private final List<Path> libraryPaths = new ArrayList<>();
     private final Path librariesFolder;
     private final boolean cleanIfNoCache;
+    private boolean loaded = false;
 
     public LibraryDownloader(Path librariesFolder) {
         this(librariesFolder, false);
@@ -36,19 +40,38 @@ public final class LibraryDownloader {
             this.deletePath(librariesFolder);
         }
         this.libraryCache.load();
+        this.addRepository(new Repository("https://repo.artillex-studios.com/releases/"));
+        this.addRepository(new Repository("https://repo1.maven.org/maven2/"));
+        this.helper = new RelocationHelper(this);
+        this.loaded = true;
     }
 
     private void deletePath(Path path) {
+        if (!Files.exists(path)) {
+            return;
+        }
+
+        LogUtils.info("Cleaning up library at path: {}", path);
         try {
+            if (Files.isDirectory(path)) {
+                Files.list(path).forEach(this::deletePath);
+            }
             Files.delete(path);
         } catch (IOException exception) {
             LogUtils.error("An exception occurred while cleaning library cache!", exception);
         }
     }
 
-    public void addLibrary(Library library) {
+    public Path addLibrary(Library library) {
+        return this.addLibrary(library, false);
+    }
+
+    public Path addLibrary(Library library, boolean transitive) {
+        long start = System.currentTimeMillis();
+        LogUtils.info("Loading library {}", library);
         if (this.isAlreadyLoaded(library)) {
-            return;
+            LogUtils.info("Library {} has already been loaded, skipping!", library);
+            return null;
         }
 
         Library cachedLibrary = this.libraryCache.getCachedLibrary(library);
@@ -57,27 +80,44 @@ public final class LibraryDownloader {
             if (FeatureFlags.DEBUG.get()) {
                 LogUtils.debug("Using cached library: {}", library);
             }
+            LogUtils.info("Loading {} from cache...", library);
             // We already have the correct version of the library cached, we can just load it
             this.libraryPaths.add(path);
             this.libraries.add(cachedLibrary);
             for (Library transitiveDependency : cachedLibrary.transitiveDependencies()) {
-                this.addLibrary(transitiveDependency);
+                this.addLibrary(transitiveDependency, true);
             }
-            return;
+            LogUtils.info("Library {} loaded in {}ms from cache!", library, System.currentTimeMillis() - start);
+            return path;
         }
 
         // remove the old version with transitive dependencies
-        this.cleanupFiles(library);
+        if (cachedLibrary != null) {
+            LogUtils.info("Library {} has an older version, cleaning up...", library);
+            this.cleanupFiles(cachedLibrary);
+        }
         // find the new library with transitive dependencies
-        Library withTransitives = this.collector.withTransitiveDependencies(library);
-        this.libraryCache.cache(withTransitives);
-        this.libraries.add(withTransitives);
+        Library withTransitives = library;
+        if (this.loaded) {
+            LogUtils.info("Searching for transitive dependencies of library: {}...", library);
+            withTransitives = this.collector.withTransitiveDependencies(library);
+            LogUtils.info("Found {} transitive dependencies of library: {}!", withTransitives.transitiveDependencies().size(), withTransitives);
+            if (!transitive) {
+                this.libraryCache.cache(withTransitives);
+            }
+        }
+
+        if (this.loaded) {
+            this.libraries.add(withTransitives);
+        }
         this.downloadLibrary(withTransitives);
+        LogUtils.info("Downloaded library {} in {}ms!", library, System.currentTimeMillis() - start);
+        return path;
     }
 
     private boolean isAlreadyLoaded(Library library) {
         for (Library loaded : this.libraries) {
-            if (this.libraryCache.checkWithoutVersion(library, loaded)) {
+            if (LibraryCache.checkWithoutVersion(library, loaded)) {
                 return true;
             }
         }
@@ -98,21 +138,30 @@ public final class LibraryDownloader {
             this.downloadLibrary(transitiveDependency);
         }
 
+        LogUtils.info("Downloading library {}...", library);
         Path path = this.getPath(library);
-        // TODO: Relocations
         for (Repository repository : this.getRepositories()) {
             URI jarURI = repository.getJarURI(library);
-            try (InputStream inputStream = jarURI.toURL().openStream(); FileOutputStream stream = new FileOutputStream(path.toFile())) {
+            Path resolve = this.loaded ? path.resolveSibling(path.getFileName() + ".resolving") : path;
+            try (InputStream inputStream = jarURI.toURL().openStream(); FileOutputStream stream = new FileOutputStream(resolve.toFile())) {
                 byte[] bytes = new byte[8192];
                 int length;
                 while ((length = inputStream.read(bytes)) != -1) {
                     stream.write(bytes, 0, length);
                 }
 
+                if (this.loaded) {
+                    LogUtils.info("Relocating library: {}", library);
+                    this.helper.relocate(resolve, path, this.createRelocationsMap());
+                    this.deletePath(resolve);
+                    LogUtils.info("Relocated library: {}", library);
+                }
                 this.libraryPaths.add(path);
             } catch (IOException exception) {
+                exception.printStackTrace();
                 LogUtils.error("An exception occurred while downloading library {}!", library, exception);
             }
+            break;
         }
     }
 
@@ -128,14 +177,26 @@ public final class LibraryDownloader {
         if (this.libraryCache.cache(relocation)) {
             return;
         }
+
+        LogUtils.info("Re-downloading libraries, as a new relocation has been added!");
         // We need to remove all cached jars and relocate them
         this.deletePath(this.librariesFolder);
         this.libraryCache.save();
+        this.collector.reset();
 
         // Download the libraries again
         for (Library library : this.libraries) {
             this.downloadLibrary(library);
         }
+    }
+
+    private Map<String, String> createRelocationsMap() {
+        Map<String, String> relocations = new HashMap<>();
+        for (Relocation relocation : this.relocations) {
+            relocations.put(relocation.from(), relocation.to());
+        }
+
+        return relocations;
     }
 
     private Path getPath(Library library) {
